break out verify tests into their own folder
