+++ packages/hardhat-plugin/test/verify/index.ts
@@ -1,0 +1,22 @@
+/* eslint-disable import/no-unused-modules */
+import { assert } from "chai";
+
+import { useEphemeralIgnitionProject } from "../use-ignition-project";
+
+describe("verify", function () {
+  describe("when there is no etherscan API key configured", function () {
+    useEphemeralIgnitionProject("verify-no-api-key");
+
+    it("should throw", async function () {
+      await assert.isRejected(
+        this.hre.run(
+          { scope: "ignition", task: "verify" },
+          {
+            deploymentId: "test",
+          }
+        ),
+        /No etherscan API key configured/
+      );
+    });
+  });
+});
